# Improve display fmt performance