--- a/src/padded.rs
+++ b/src/padded.rs
@@ -8,7 +8,7 @@ pub struct Z85p {
 
 impl fmt::Display for Z85p {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{}", self.as_str())
+        f.write_str(self.as_str())
     }
 }