Interestingly, the `write!(f, "{}", self.as_str())` call doesn't optimize as well as I would've thought, so one needs to be more explicit to get good performance out of the `Display` impl.

Running the included benchmarks, this change shows a good amount of improvement (~10% reduction in runtime) on my machine:

```
rfc                     time:   [358.13 ns 364.20 ns 372.14 ns]
                        change: [-12.932% -11.637% -10.265%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  1 (1.00%) high mild
  3 (3.00%) high severe

padded                  time:   [363.56 ns 365.61 ns 367.97 ns]
                        change: [-11.080% -9.8406% -8.5790%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  4 (4.00%) high mild
  3 (3.00%) high severe
```

The benchmark is not super clean, because it includes `.clone` and `format!`, both of which will likely allocate, so the actual improvement for the `fmt` method alone may be more significant that indicated by the benchmark results.